[DEVELOPMENT ENVIRONMENT] * NEW: You can open a terminal in the project directory now. * NEW: The IDE now marks projects as Gambas 3 projects. * NEW: Port Gambas2 fixes.
(INTERPRETER]
* NEW: New API for returning the type of a Gambas array.

[GB.DESKTOP]
* NEW: New internal methods for reading or writing a windows property, and
  for sending a client message to the root window.
* NEW: Desktop.Find() has been renamed to Desktop.FindWindow().
* NEW: Desktop.OpenTerminal() opens a terminal on a specified directory 
  according to the current desktop in use.

[GB.QT]
* BUG: Window id are really integers, not pointers.

[GB.GTK]
* BUG: Window id are really integers, not pointers.


git-svn-id: svn://localhost/gambas/trunk@1276 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ gb.gtk/src/CWindow.cpp
@@ -37,7 +37,7 @@
 #include "CDraw.h"
 #include "gapplication.h"
 
-long CWINDOW_Embedder = 0;
+int CWINDOW_Embedder = 0;
 bool CWINDOW_Embedded = false;
 
 CWINDOW *CWINDOW_Active = NULL;
